Add: handlers folder and moved some files there
